+++ lib/company_analytics.dart
@@ -2,4 +2,5 @@ export 'src/analytics_config.dart';
 export 'src/analytics_event.dart';
 export 'src/analytics_exception.dart';
 export 'src/company_analytics.dart';
+export 'src/remote_analytics_config.dart';
 export 'src/sdk_singletons.dart';